feat: a lot of changes; see commit description
Before you get scared of the size of this diff, let me explain.

Doctave (from which Docgen has been forked) is a mess. I say "is"
because I haven't yet done a rewrite. I say that knowing the
trade-offs that come with a rewrite. I spent the last couple of weeks
trying to bring "performance" to the codebase. In my benchmarks
Docgen should be 2x or more faster. I'll post proper benchmarks soon.

Here are the list of changes (I probably should have commited them
separately but I got carried away):

1. I got rid of the whole "Directory" concept which was unnecessary
and required nested where we could do with plain vectors.
2. A lot of unnecessary memory allocations have been removed. We
do less cloning now & more references. This also
simplified the codebase.
3. Handlebars has been gotten rid of and replaced with markup. Why?
Because Handlebars is slow. Markup also brings with it Rust safety
so yay!
4. katex is now an optional feature. We use `latex2mathml` by default.
It isn't as good as `katex` but its native rust and hence much faster.
5. Docgen sites are now fully themeable. All the colors are basically
css variables. It is a mess right now though.
6. And many other small features that I'll find and list in the release
notes.

Docgen is not, by any means, at peak performance. Transforming md to
html shouldn't be this slow. But it's good enough for now.